requirements.txt
boto3==1.18.63
botocore==1.21.63
jmespath==0.10.0
python-dateutil==2.8.2
python-dotenv==0.19.1
pkg_resources==0.0.0
RPi.GPIO==0.7.0
s3transfer==0.5.0
six==1.16.0
urllib3==1.26.7
